Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> i think LICENSE should be NOTICE and just be a symlink to COPYING.
> 
> there should be a MODULE_LICENSE_BSD.
> 
> as for METADATA... does that replace README.version? we still have a lot of README.version and almost no METADATA files (though this wouldn't be the first). what uses METADATA?

Changing the LICENSE to a symlink and add NOTICE as a symlink to the same thing. The new process had me create a LICENSE file, but no NOTICE file.

The new process for adding third-party code has a step to generate the METADATA file. I don't know who uses it.

There is no README.version in the old jemalloc, I don't know when that was required.

Patch-set: 1
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>