Remove best fit for extent selection.
A jemalloc user reported that the best fit selection is causing them
a memory leak. This code has been completely removed from the next
release of jemalloc (5.2.0), so remove it since it doesn't have any
real benefit.

Running the memory dumps, removing best fit appears to be a win: it is
slightly faster and has the same PSS/VA.

Bug: 128697497

Test: Ran jemalloc unit tests.
Test: Ran memory dumps in 32 bit and 64 bit and observed that the PSS
Test: and VA stayed the same, while run time improved slightly.
Change-Id: I98a8ddf2cea837c8ade1afd4a998960c253d3932